Reorganize widgets under `widgets` directory
