+++ src/widgets/mod.rs
@@ -1,0 +1,4 @@
+pub mod split;
+pub mod dualslider;
+pub mod toggler;
+pub mod viewer;
